+++ agent/mibgroup/ip-mib/ipIfStatsTable/ipIfStatsTable_data_access.c
@@ -346,7 +346,7 @@ ipIfStatsTable_container_load(netsnmp_container * container)
                                           NETSNMP_ACCESS_SYSTEMSTATS_FREE_DONT_CLEAR);
 
     DEBUGMSGT(("verbose:ipIfStatsTable:ipIfStatsTable_container_load",
-               "%d records\n", CONTAINER_SIZE(container)));
+               "%lu records\n", (unsigned long)CONTAINER_SIZE(container)));
 
     return MFD_SUCCESS;
 }                               /* ipIfStatsTable_container_load */
